Move 'MapDaisy' into 'Microsoft.AspNetCore.Builder' namespace
